refactor: Move handler and change config
