+++ central/main.go
@@ -107,11 +107,11 @@ import (
 	"github.com/stackrox/rox/central/pruning"
 	rbacService "github.com/stackrox/rox/central/rbac/service"
 	reportConfigurationService "github.com/stackrox/rox/central/reports/config/service"
-	reportHandler "github.com/stackrox/rox/central/reports/handler"
 	vulnReportScheduleManager "github.com/stackrox/rox/central/reports/manager"
 	vulnReportV2Scheduler "github.com/stackrox/rox/central/reports/scheduler/v2"
 	reportService "github.com/stackrox/rox/central/reports/service"
 	reportServiceV2 "github.com/stackrox/rox/central/reports/service/v2"
+	v2Service "github.com/stackrox/rox/central/reports/service/v2"
 	"github.com/stackrox/rox/central/reprocessor"
 	collectionService "github.com/stackrox/rox/central/resourcecollection/service"
 	"github.com/stackrox/rox/central/risk/handlers/timeline"
@@ -777,8 +777,13 @@ func customRoutes() (customRoutes []routes.CustomRoute) {
 	)
 
 	if env.VulnReportingEnhancements.BooleanSetting() {
-		// Report custom routes
-		customRoutes = append(customRoutes, reportHandler.CustomRoutes()...)
+		// Append report custom routes
+		customRoutes = append(customRoutes, routes.CustomRoute{
+			Route:         "/api/reports/jobs/download",
+			Authorizer:    user.With(permissions.Modify(resources.WorkflowAdministration)),
+			ServerHandler: v2Service.NewDownloadHandler(),
+			Compression:   true,
+		})
 	}
 
 	debugRoutes := utils.IfThenElse(env.ManagedCentral.BooleanSetting(), []routes.CustomRoute{}, debugRoutes())
