+++ central/reports/service/v2/download_test.go
@@ -1,4 +1,4 @@
-package handler
+package v2
 
 import (
 	"context"
