+++ proto/storage/config.proto
@@ -57,7 +57,7 @@ message DecommissionedClusterRetentionConfig {
 
 // next available tag: 4
 message ReportRetentionConfig {
-    int32 history_retention_duration_days              = 1;
+    uint32 history_retention_duration_days              = 1;
     uint32 downloadable_report_retention_days          = 2;
     uint32 downloadable_report_global_retention_bytes  = 3;
 }
